Reorganize all files in the directory
