+++ src/app/app.scss
@@ -1,4 +1,1 @@
 @import '~react-components/styles/index';
-
-@import './components/message/styles/embedded.scss';
-@import './components/message/styles/MessageBody.scss';
