+++ src/app/components/message/MessageBody.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { isPlainText } from './logic/message';
+import { isPlainText } from '../../helpers/message';
 import { classnames } from 'react-components';
+
+import './MessageBody.scss';
 
 const MessageBody = ({ message: { content, data: message } }) => {
     const plain = isPlainText(message);
