+++ src/app/components/message/MessageBody.scss
@@ -1,10 +1,36 @@
+.bodyDecrypted {
+    /* Forbidden property on almost ALL clients IE add a space :) */
+    [style*='position:absolute'],
+    [style*='position: absolute'] {
+        position: relative !important;
+    }
+
+    table {
+        table-layout: auto;
+    }
+
+    // blockquote {
+    //     @extend [class*='block-info'], [class*='block-info-standard'], .block-info-standard;
+    // }
+}
+
+.bodyDecrypted.plain {
+    white-space: pre-wrap;
+    font-family: monospace;
+}
+
+.bodyDecrypted pre code {
+    white-space: pre-wrap;
+    background-color: transparent;
+}
+
 /* Also exist inside squireEditor.js */
 .proton-embedded:not([src]) {
     position: relative;
     min-height: 38px; /* placeholder */
     border: 1px solid;
     border-color: #444 #ccc #ccc #444;
-    background: url('./broken-img.png') no-repeat 0 50% $white;
+    background: url('../../assets/broken-img.png') no-repeat 0 50% #fff;
 }
 
 /** safari can't override the img border on non-src img
@@ -40,7 +66,7 @@
     text-overflow: ellipsis;
     padding: 10px 0 0 20px;
     color: rgba(0, 0, 0, 0.5);
-    background: url('./broken-img.png') no-repeat 0 50% $white; /* override the background in far left */
+    background: url('../../assets/broken-img.png') no-repeat 0 50% #fff; /* override the background in far left */
 }
 
 /* Responsiveness of our embedded images */
