--- src/app/components/message/styles/MessageBody.scss
@@ -1,25 +1,0 @@
-.bodyDecrypted {
-    /* Forbidden property on almost ALL clients IE add a space :) */
-    [style*='position:absolute'],
-    [style*='position: absolute'] {
-        position: relative !important;
-    }
-
-    table {
-        table-layout: auto;
-    }
-
-    blockquote {
-        // @extend [class*='block-info'], [class*='block-info-standard'], .block-info-standard;
-    }
-}
-
-.bodyDecrypted.plain {
-    white-space: pre-wrap;
-    font-family: monospace;
-}
-
-.bodyDecrypted pre code {
-    white-space: pre-wrap;
-    background-color: transparent;
-}
