+++ src/app/helpers/displaySignature.js
