+++ src/app/helpers/embedded/embeddedFinder.js
