+++ src/app/helpers/embedded/embeddedStore.js
