+++ src/app/helpers/encryptionType.js
