+++ src/app/helpers/signatures.js
