+++ src/app/helpers/string.js
