+++ src/app/helpers/transforms/transformEscape.js
@@ -1,6 +1,6 @@
-import { uniqID } from '../../helpers/stringHelper';
-import { protonizer as purifyHTML } from '../../helpers/purifyHelper';
-import { parseInDiv } from '../../helpers/domHelper';
+import { uniqID } from '../string';
+import { protonizer as purifyHTML } from '../purify';
+import { parseInDiv } from '../dom';
 
 /*
  * match attributes or elements with svg, xlink, srcset, src, background, poster.
