+++ src/app/helpers/transforms/transformWelcome.js
