+++ src/app/hooks/useComputeMessage.js
@@ -1,13 +1,13 @@
 import { useCallback } from 'react';
 import { useCache } from 'react-components';
-import { transformEscape } from '../logic/transforms/transformEscape';
-import { transformLinks } from '../logic/transforms/transformLinks';
-import { transformEmbedded } from '../logic/transforms/transformEmbedded';
-import { transformWelcome } from '../logic/transforms/transformWelcome';
-import { transformBlockquotes } from '../logic/transforms/transformBlockquotes';
-import { transformStylesheet } from '../logic/transforms/transformStylesheet';
-import { transformRemote } from '../logic/transforms/transformRemote';
-import { transformBase } from '../logic/transforms/transformBase';
+import { transformEscape } from '../helpers/transforms/transformEscape';
+import { transformLinks } from '../helpers/transforms/transformLinks';
+import { transformEmbedded } from '../helpers/transforms/transformEmbedded';
+import { transformWelcome } from '../helpers/transforms/transformWelcome';
+import { transformBlockquotes } from '../helpers/transforms/transformBlockquotes';
+import { transformStylesheet } from '../helpers/transforms/transformStylesheet';
+import { transformRemote } from '../helpers/transforms/transformRemote';
+import { transformBase } from '../helpers/transforms/transformBase';
 import { useDecryptMessage } from './useDecryptMessage';
 import { useLoadMessage } from './useLoadMessage';
 import { useMarkAsRead } from './useMarkAsRead';
