+++ src/app/hooks/useSignatures.js
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import { useCache } from 'react-components';
-import { verify as verifySignature } from '../logic/signatures';
+import { verify as verifySignature } from '../helpers/signatures';
 
 const CACHE_KEY = 'Signatures';
